Switch to using a first class object for qasino tables. Qasino table objects have properties associated with them to govern things like if its a static table, persistence, and whether to do an "update" vs. the standard insert.
Also support optionally setting properties in the first line
of csv files read by qasino_csvpublisher.  This way publishers
can specify if they want the table to be static, persist, update
etc.

Lastly support a more specialized update semantic where a
list of key columns is provided.  Then use an INSERT OR REPLACE
command (after a unique constraint is applied to the columns)
to update (or insert) rows based on the key columns.